Add separate logger package
