+++ common/changes/just-task/ecraig-logger_2019-02-21-23-18.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "packageName": "just-task",
+      "comment": "Use new logger package",
+      "type": "patch"
+    }
+  ],
+  "packageName": "just-task",
+  "email": "elcraig@microsoft.com"
+}
