+++ packages/just-scripts-utils/package.json
@@ -13,11 +13,11 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "chalk": "^2.4.1",
     "fs-extra": "^7.0.1",
     "glob": "^7.1.3",
     "handlebars": "^4.0.12",
     "jju": "^1.4.0",
+    "just-task-logger": ">=0.1.0 <1.0.0",
     "marked": "^0.6.0",
     "marked-terminal": "^3.2.0",
     "semver": "^5.6.0",
