+++ packages/just-task-logger/.npmignore
@@ -1,0 +1,3 @@
+src
+CHANGELOG.*
+__tests__
