+++ packages/just-task-logger/jest.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  roots: ['<rootDir>/src'],
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  testMatch: ['**/?(*.)+(spec|test).ts']
+};
