+++ packages/just-task-logger/src/__tests__/__mocks__/yargs.ts
@@ -1,0 +1,6 @@
+import { Arguments } from 'yargs';
+
+const yargs = jest.genMockFromModule<Arguments>('yargs');
+yargs.argv = {};
+
+module.exports = yargs;
