+++ packages/just-task-logger/src/index.ts
@@ -1,0 +1,1 @@
+export { logger, ILogger } from './logger';
