+++ packages/just-task-logger/tsconfig.json
@@ -1,0 +1,14 @@
+{
+  "compilerOptions": {
+    "target": "es2015",
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "declaration": true,
+    "declarationMap": true,
+    "outDir": "lib",
+    "strict": true,
+    "strictBindCallApply": false,
+    "esModuleInterop": true
+  },
+  "include": ["src"]
+}
