+++ packages/just-task/package.json
@@ -14,9 +14,10 @@
   },
   "dependencies": {
     "chalk": "^2.4.1",
+    "just-task-logger": ">=0.1.0 <1.0.0",
+    "resolve": "^1.8.1",
     "undertaker": "^1.2.0",
     "undertaker-registry": "^1.0.1",
-    "resolve": "^1.8.1",
     "yargs": "^12.0.5"
   },
   "devDependencies": {
