+++ packages/just-task/src/logger.ts
@@ -1,28 +1,1 @@
-import chalk from 'chalk';
-
-let longestTaskNameLength: number = 0;
-
-function getTimestamp() {
-  const now = new Date();
-  return `[${now.toLocaleTimeString()}]`;
-}
-
-export interface ILogger {
-  info(msg?: any, ...optionalParams: any[]): void;
-  warn(msg?: any, ...optionalParams: any[]): void;
-  error(msg?: any, ...optionalParams: any[]): void;
-}
-
-export const logger: ILogger = {
-  info(msg?: any, ...optionalParams: any[]) {
-    console.info.apply(null, [`${chalk.gray(getTimestamp())} ${chalk.green('\u25a0')} ${msg}`, ...optionalParams]);
-  },
-
-  warn(msg?: any, ...optionalParams: any[]) {
-    console.warn.apply(null, [`${chalk.gray(getTimestamp())} ${chalk.yellow('\u25b2')} ${msg}`, ...optionalParams]);
-  },
-
-  error(msg?: any, ...optionalParams: any[]) {
-    console.error.apply(null, [`${chalk.gray(getTimestamp())} ${chalk.redBright('x')} ${msg}`, ...optionalParams]);
-  }
-};
+export * from 'just-task-logger';
