+++ rush.json
@@ -106,6 +106,11 @@
       "shouldPublish": false
     },
     {
+      "packageName": "just-task-logger",
+      "projectFolder": "packages/just-task-logger",
+      "shouldPublish": true
+    },
+    {
       "packageName": "example-lib",
       "projectFolder": "packages/example-lib",
       "shouldPublish": false
